refactor(agent): 全面架构优化 — 线程池化/流式LLM/注册表模式/安全加固/审计日志
P0 (紧急):
- 线程池化: AgentExecutorConfig 替代 new Thread()，core=4/max=16/queue=100
- LLM 流式响应: chatStream() 逐 token SSE 推送，首字节延迟降低
- API Key 脱敏: 仅暴露后4位，前缀替换为固定值

P1 (代码质量):
- AgentResult<T> + AgentException 异常体系，消除 null 返回
- LuneApiClient 返回 Optional 替代 null
- ToolExecutor 注册表模式重构 (10个领域Handler)
- LLMClient 指数退避重试 (3次)

P2 (性能/资源):
- Dashboard 并行 CompletableFuture 统计查询
- ToolExecutor 30s 超时保护
- 优雅关闭 + Redis 连接池 + GC 日志/OOM dump
- application.yml 配置对齐

P3 (安全/运维):
- Prompt injection 防御 (安全规则 + XML 标签包裹)
- ChatMemory 多 Session 支持
- AuditLogger 工具调用审计日志

+++ lune-server/lune-web/src/main/java/com/lune/common/PageResult.java
@@ -7,10 +7,10 @@ import java.util.List;
 public class PageResult<T> {
     private List<T> records;
     private long total;
-    private long page;
-    private long size;
+    private int page;
+    private int size;
 
-    public static <T> PageResult<T> of(List<T> records, long total, long page, long size) {
+    public static <T> PageResult<T> of(List<T> records, long total, int page, int size) {
         PageResult<T> result = new PageResult<>();
         result.setRecords(records);
         result.setTotal(total);
